reorganize div sections and stuff
