+++ apps/interim/components/email-newsletter-input-section/email-newsletter-input-section.spec.tsx
@@ -1,0 +1,10 @@
+import { render } from '@testing-library/react';
+
+import EmailNewsletterInputSection from './email-newsletter-input-section';
+
+describe('EmailNewsletterInputSection', () => {
+  it('should render successfully', () => {
+    const { baseElement } = render(<EmailNewsletterInputSection />);
+    expect(baseElement).toBeTruthy();
+  });
+});
