+++ apps/interim/components/logo-and-slogan/logo-and-slogan.spec.tsx
@@ -1,0 +1,10 @@
+import { render } from '@testing-library/react';
+
+import LogoAndSloganSection from './logo-and-slogan';
+
+describe('LogoAndSlogan', () => {
+  it('should render successfully', () => {
+    const { baseElement } = render(<LogoAndSloganSection />);
+    expect(baseElement).toBeTruthy();
+  });
+});
